Membuat hapus bberapa dgn livewire

+++ app/Livewire/Employee.php
@@ -17,6 +17,7 @@ class Employee extends Component
     public $updateData = false;
     public $employee_id;
     public $katakunci;
+    public $employee_selected_id = [];
 
 
     public function store()
@@ -82,34 +83,43 @@ class Employee extends Component
 
         $this->updateData = false;
         $this->employee_id = '';
+        $this->employee_selected_id = [];
     }
 
     public function delete()
     {
-        $id = $this->employee_id;
-        ModelsEmployee::find($id)->delete();
-
+        if ($this->employee_id != '') {
+            $id = $this->employee_id;
+            ModelsEmployee::find($id)->delete();
+        }
+        if (count($this->employee_selected_id)) {
+            for($x=0;$x<count($this->employee_selected_id);$x++){
+                ModelsEmployee::find($this->employee_selected_id[$x])->delete();
+            }
+        }
         session()->flash('message', 'Data berhasil dihapus');
     }
 
     public function delete_konfirmasi($id)
     {
-        $this->employee_id = $id;
+        if ($id != '') {
+            $this->employee_id = $id;
+        }
     }
 
-    
+
     public function render()
     {
-        if ($this->katakunci != null){
-            $data = ModelsEmployee::where('nama','like', '%' . $this->katakunci . '%')
-            ->orWhere('email','like', '%' . $this->katakunci . '%')
-            ->orWhere('alamat','like', '%' . $this->katakunci . '%')
-            ->orderBy('nama', 'asc')->paginate(2);
-        }else{
+        if ($this->katakunci != null) {
+            $data = ModelsEmployee::where('nama', 'like', '%' . $this->katakunci . '%')
+                ->orWhere('email', 'like', '%' . $this->katakunci . '%')
+                ->orWhere('alamat', 'like', '%' . $this->katakunci . '%')
+                ->orderBy('nama', 'asc')->paginate(2);
+        } else {
             $data = ModelsEmployee::orderBy('nama', 'asc')->paginate(2);
         }
 
-        
+
         return view('livewire.employee', ['dataEmployees' => $data]);
     }
 }
